Added the chronological navigation to the post page.
While following Nick's post series I felt the need for easy navigation from
current post to previous and next post in the series, thus avoiding going
back to the index page again and again.

Also it is helpful for somebody coming in from a search engine to read the
posts in order, or for archive pages.

This branch can be safely merged with the delete_handler branch and is kept
separate just as a choice to the user. The delete takes care of adjusting the
previous and next links automatically.